requirements.txt


# requirements.txt

streamlit>=1.28.0
pymongo>=4.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-mongodb>=0.1.0
langchain-google-genai>=0.0.6
pypdf>=3.17.0
sentence-transformers>=2.2.2
gpt4all>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.17.0
streamlit-option-menu>=0.3.6

